Initial Commit
Finished the basic framework of the mod.
TO DO: textures and stats for other blocks
TO DO: config file to either customize block attributes or add any block
TO DO: decide on crafting recipe. 3x3 with pumpkin, book, and redstone,
or 2-step process with a golem spell (book and redstone) and pumpkin?
TO DO: other code cleanup and lang file