AI commit
Added several AIs that are triggered when using certain blocks for the
golem.
Working: EntityAIPotionBuffs, EntityAIProvideLight,
EntityAIPlaceRandomBlocks, EntityAITeleportRandomly
Buggy: EntityAIExplode, EntityAIProvidePower
Added stats for most of the blocks I have textures for (and added many
textures). Trying to get the mappings to return the default state (when
given any IBlockState) unless I specify otherwise, but I still want to
be able to specify otherwise.
Added GolemBuildEvent for who knows what reason... I was going to use it
for adding AIs but made an AIManager class instead.